Fix clippy warnings in core, parser, and rendering modules
Pure refactors, no behavior change: div_ceil/is_multiple_of intrinsics,
let-chains for collapsible ifs, strip_prefix/copy_from_slice, slice::from_ref,
derive(Default) for Color/RenderQuality, if-let rewrites in input.rs key
encoders, and doc-comment escaping in parser/machine.rs. too_many_arguments and
large_enum_variant resolved via scoped allow-attributes per existing convention.

+++ src/emoji/color_atlas.rs
@@ -181,7 +181,7 @@ impl ColorGlyphAtlas {
         let (x0, y0) = self.slot_origin(slot.slot);
         let row_bytes = slot.width_cells as usize * self.cell.width as usize * 4;
         for row in 0..self.cell.height as usize {
-            let dst = (((y0 as usize + row) * self.width as usize + x0 as usize) * 4) as usize;
+            let dst = ((y0 as usize + row) * self.width as usize + x0 as usize) * 4;
             let src = row * row_bytes;
             self.data[dst..dst + row_bytes].copy_from_slice(&rgba[src..src + row_bytes]);
         }
